fix(api): set serial timeouts for transactions
The async-serial refactor changed serial timeouts to be set for
individual serial actions (aka a single call to serial.write or
serial.read_until), rather than a write-read transaction pair as it
previously had been. While this shouldn't be a problem, and isn't a
problem on the OT-2, OSX, or desktop Linux, it is in fact a problem on
Windows: https://github.com/pyserial/pyserial/issues/394

The tl;dr is that the pyserial win32 implementation does a full port
reconfigure when you change a timeout. On specifically FTDI devices, as
with test engineering running an OT-2 test fixture by connecting the
motor controller to a laptop with an FTDI cable and running this
software on the laptop, a port reconfigure during an ongoing serial
transaction - which the host of course doesn't really control - can and
sometimes will corrupt data.

The solution to this is to move timeout setting back to a transaction
timeframe, so that there are overall fewer timeout changes and thus
fewer port reconfigures and thus the problem doesn't occur.