refactor(api,robot-server): generic ThreadManager
The ThreadManager and SynchronousAdapter classes wrap... something, and
we can now specify that something.

For now this just adds some more verbosity since those classes are
implemented with getattr and therefore won't actually fail typecheck if
they wrap the wrong thing, but if we can fix that then we'll be golden.