refactor(robot-server): correct hardware typing
With the new generic thread manager and the hardware control protocol,
it's now possible to spread the love and joy of correct typing of the
hardware controller into the robot server! By using the protocol and
using the .wrapped() method that converts a generic thread manager to
whatever it wraps, all of robot server can now know if it's using the
hardware controller wrong and we'll never write a bug again.

Most of this doesn't involve any code changes. The only exceptions are
the calibration flows, which  poke into hardware controller internals
to get the internal pipette representation, which of course wasn't in
the protocols, and is now. Please don't use this anymore.