refactor(api): OT3 HC API and HC API protocol
Some changes to the OT3 hardware control strategy will not be
adequately representable at the controller level alone (for instance,
switching the direction the pipette plunger moves during liquid
handling). In addition, the HCAPI<->controller API isn't really suitable
for the OT3 (for instance, in the OT3 there's no place where it's a good
idea to be using strings to represent motion axes). The
hardware control API is also pretty componentized - core system
behaviors like execution control are already separate classes with a
has-a relationship. Finally, changes at higher levels of the stack make
the value of a simulation/control split much less apparent, and with a
separate OT3 hardware control API it will be easier to fuse controller
and API from the beginning.

For these reasons it's a much better fit to make an OT3 hardware control
API than it is to make an OT3 controller.

Now that there will be two APIs, we need a nice type that represents them
both and is also compatible with the various adapters that implement
their wrapping with __getitem__. The only way to do this is with a mypy
protocol, which is compatible with __getitem__ in a way that abc is
not (__getitem__ doesn't count as implementing @abstractmethod and so
you can't inherit from the ABC in e.g. ThreadManager), and is arguably
easier to maintain than a type alias for Union[API, OT3API,
ThreadManager, ...].

This commit just establishes the protocol and changes callers to use it.