perf(api): bring back the ThreadManager (#9199)
Until we move the HardwareAPI to a child process, the HTTP API
is too heavyweight to occupy the same thread as hardware control
without interference. By bringing back the ThreadManager, we
ensure there is a hardware-only-thread in use to avoid complete
blockage - the GIL is still in play, but that is best resolved
with processes.