refactor(api,robot-server): type the threadmanager (#9201)
Some changes to the OT3 hardware control strategy will not be
adequately representable at the controller level alone (for instance,
switching the direction the pipette plunger moves during liquid
handling). In addition, the HCAPI<->controller API isn't really suitable
for the OT3 (for instance, in the OT3 there's no place where it's a good
idea to be using strings to represent motion axes). The
hardware control API is also pretty componentized - core system
behaviors like execution control are already separate classes with a
has-a relationship. Finally, changes at higher levels of the stack make
the value of a simulation/control split much less apparent, and with a
separate OT3 hardware control API it will be easier to fuse controller
and API from the beginning.

For these reasons it's a much better fit to make an OT3 hardware control
API than it is to make an OT3 controller.

Now that there will be two APIs, we need a nice type that represents them
both and is also compatible with the various adapters that implement
their wrapping with __getitem__. The only way to do this is with a mypy
protocol, which is compatible with __getitem__ in a way that abc is
not (__getitem__ doesn't count as implementing @abstractmethod and so
you can't inherit from the ABC in e.g. ThreadManager), and is arguably
easier to maintain than a type alias for Union[API, OT3API,
ThreadManager, ...].

This commit just establishes the protocol and changes callers to use it.

* refactor(api): factor out BaseHardwareControl

BaseHardwareControl as a protocol really fuses together a set of
otherwise unrelated or semirelated functionality. By splitting it into
smaller protocols, which compose through "inheritance", we can then
split this large file up and change callers to require only the
functionality they need.

* refactor(api): split protocols into package

Bit more readable, no more giant file, better names for things

* refactor(api,robot-server): generic ThreadManager

The ThreadManager and SynchronousAdapter classes wrap... something, and
we can now specify that something.

For now this just adds some more verbosity since those classes are
implemented with getattr and therefore won't actually fail typecheck if
they wrap the wrong thing, but if we can fix that then we'll be golden.

* refactor(hardware_control): fix up typing nits

Mypy doesn't really like implicit optionals when using protocols so make
them explicit.

* refactor(robot-server): correct hardware typing

With the new generic thread manager and the hardware control protocol,
it's now possible to spread the love and joy of correct typing of the
hardware controller into the robot server! By using the protocol and
using the .wrapped() method that converts a generic thread manager to
whatever it wraps, all of robot server can now know if it's using the
hardware controller wrong and we'll never write a bug again.

Most of this doesn't involve any code changes. The only exceptions are
the calibration flows, which  poke into hardware controller internals
to get the internal pipette representation, which of course wasn't in
the protocols, and is now. Please don't use this anymore.


Co-authored-by: Mike Cousins <mike@opentrons.com>