fix(api): Prioritize mc alarms over errors (#9230)
* fix(api): Prioritize mc alarms over errors

The error message returned by the motor controller after a hard halt is
"error:Alarm lock". This is meant to be sensed by logic that raises a
different kind of exception (SmoothieAlarm) from the kind raised by for
instance an unexpected limit switch actuation or malformed
response (SmoothieError). Unfortunately, the _way_ this is sensed is
that SmoothieAlarm looks for the keyword "alarm" and SmoothieError looks
for the keyword "error", and that message contains both, and the error
case is checked first.

The fix is to swap the order of the two. The fix could also be to
special case the message checking, but let's see if keeping it simple
works first.

* Clear hard halting in reset from error

Should prevent a latched hard-halting condition from occurring that
would incorrectly not transform limit switch alarms to errors.

* add tests for halt behavior