refactor(api,shared-data,robot-server):  ot3 deck (#9207)
* refactor(shared-data): add ot3_standard deck

Adds a definition for the OT-3 standard deck (and the trash labware).
This will likely not be the final definition - it will need to grow more
complex since the trash, for instance, is likely movable, and also
prototypes will change - but it is accurate for and tested on current
hardware.

Co-Developed-By: Laura Cox <laura@opentrons.com>

* refactor(api): add function to determine deck

The type of deck used on the current robot is currently determined
piecemeal around the system by every piece of code that needs to know
the deck. With the deck beginning to change more frequently and
fundamentally, it will be useful to centralize these places. A new
function in the same place that stores the standard deck can accomplish
this.

Co-Developed-By: Laura Cox <laura@opentrons.com>

* refactor(api,robot-server): use deck from FF

The type of deck is now primarily determined by the value of the ot3
feature flag (and additionally whether the short-trash flag is on). This
uses the new centralized determination of deck type.

Co-Developed-By: Laura Cox <laura@opentrons.com>

* update robot version description

* Update api/tests/opentrons/protocol_engine/resources/test_deck_data_provider.py

Co-authored-by: Mike Cousins <mike@opentrons.com>