refactor(api): Test ot3 controller
By adapting the hardware fixture to provide both ot2 and ot3 apis, we
get free regression testing! There's a couple functions that outright
won't work on the other one, and a mark is provided for that. A lot of
tests had to be fixed up to mock methods of the injected hardware object
rather than the class, and I took the time to switch them to use
mock.patch which I like way more than pytest monkeypatch fixture.