refactor(api): centralize pickup behavior
Move pick up tip behavior at a high level to the instrument handler.
This is a particularly intense refactor because the pickup tip behavior
is both a) complicated and b) split up between multiple backend calls
which remain in the child class.

As with the other liquid handler centralizations, the way around this is
to make a fairly big frozen dataclass containing the arguments (or
things trivially convertible to the arguments) to the backend calls, and
the functionality is confirmed to be regression free by relying on tests.