perf(app): virtualize command list on run detail page
In order to make long protocols usable in the new UI, we should only render a small window of the
entire commands list to cover what users are actually able to see in the viewport. Previously large
protocols would gum up the js prcoess, but this should allow protocols of all sizes to behave like
protocols with only 100 steps in terms of virtual DOM performanc. also adds numbers to each command
item which is very useful for visually anchoring users of long protocols

Closes #9217