refactor(api): ot3 hardware control API (#9062)
Switches the OT3 hardware controller to a whole new hardware control API rather than just a controller.

Some changes required for the OT3 to work - the big one is that the plunger runs in the other direction - really can't be entirely handled in the controller layer alone, and need hardware API changes. With #9201 , it's much more reasonable to have two hardware control APIs with the same interface and working with typechecking, so this PR creates that second API and adds the changes required for the OT3 to run protocols.

It also adds testing for the new API, by extending the hardware fixture in the API tests to provide both ot3 and ot2 hardware controllers. To me, this testing is sufficient to not have to do the same thing in the robot server, but I'm interested in other opinions.

The next steps after this API is to start pulling common implementation into shared code rather than leaving it duplicated between the two APIs, and either merge the controller code into the API or (since we may need the Simulation layer until we can finally get rid of api-level simulation) change the interface to better-reflect the internal data structures of the OT3 hardware controller.

This PR when merged should allow protocols to be run on the OT3 as long as you're ok with not homing, having hardcoded pipettes, and changing the default deck calibration to [[1, 0, 0], [0, -1, 0], [0, 0, -1]] and the pipette calibration to have different offsets. Those defaults will follow in other PRs.

Co-authored-by: Laura Cox <lauracox@Lauras-MacBook-Pro.local>