refactor(api): Separate OT3 robot config
The OT3 has a very different set of required settings from the OT2.
Create a new dataclass and loading/saving functions to hold them and
propagate types around.

The extra settings here are meant for axis constraints for the motion
control system. They have a layered format. The "none" options apply
when nothing is connected, but as more mass is added to the head we will
need different configuration values. Values for the appropriate hardware
config should be overlaid, as in pipettes.py.

The Z axis is specified just once because it should apply equally to
both. The P axis stands for whichever plunger is active.