refactor(api): Separate OT3 robot config (#9319)
The OT3 has a very different set of required settings from the OT2.
Create a new dataclass and loading/saving functions to hold them and
propagate types around.

The extra settings here are meant for axis constraints for the motion
control system. They have a layered format. The "none" options apply
when nothing is connected, but as more mass is added to the head we will
need different configuration values. Values for the appropriate hardware
config should be overlaid, as in pipettes.py.

The Z axis is specified just once because it should apply equally to
both. The P axis stands for whichever plunger is active.

The ot3 can't actually use identity as a default deck calibration, it
needs some axis flips. The defaults are applied in
hardware_control.robot_calibration so change them there.