refactor(api): ot3-specific simulator backend
This backend is necessary because we have different home positions among
other things. We make a new submodule for backends, move them into it,
copy the ot3controller, and modify it to work more like the simulator,
then use the ot3simulator instead of the generic simulator.

There's a lot of copied code here; the next move is probably to
commonize some of it.