refactor(api): Use ot-3 transforms on ot3
This transform set has a large fixed offset to move from deck to
carriage and then an independent offset for each mount from there. Using
this type of transform set requires some small changes to account for
having an offset for both sides (the ot2 specifies the offset as (0,0,0)
for the right mount) and needs to add the carriage offset into the
machine coordinates after transformation.

There's also some annoying backend side refactors to account for not yet
having a difference between right and left pipette node IDs that I made
as annoying as possible to read and deal with so we can't forget about
them.