refactor(api): OT3 motion transforms (#9334)
Add and use OT-3 specific motion transforms between deck and machine coordinates. 

This transform set has a large fixed offset to move from deck to
carriage and then an independent offset for each mount from there. Using
this type of transform set requires some small changes to account for
having an offset for both sides (the ot2 specifies the offset as (0,0,0)
for the right mount) and needs to add the carriage offset into the
machine coordinates after transformation.

There's also some annoying backend side refactors to account for not yet
having a difference between right and left pipette node IDs that I made
as annoying as possible to read and deal with so we can't forget about
them.

Adds a new ot3simulator backend and moves the backends to a submodule.
This is necessary because we have different home positions among
other things. We make a new submodule for backends, move them into it,
copy the ot3controller, and modify it to work more like the simulator,
then use the ot3simulator instead of the generic simulator.

There's a lot of copied code here; the next move is probably to
commonize some of it.