refactor(protocol-engine): Merge APIv2 equipment brokers (#9384)
This refactors the way that APIv2 reports labware loads, pipette loads, and module loads to Protocol Engine.

Before this PR, these were reported through three separate message brokers that could be individually subscribed to and unsubscribed from. This flexibility was unnecessary, and it caused some bloat in the tests.

With this PR, these are all merged into a single `EquipmentBroker`, which brokers a single `LoadInfo` type—a union of `LabwareLoadInfo`, `InstrumentLoadInfo`, or `ModuleLoadInfo`. It's now `LegacyCommandMapper`'s job to unwrap these generic `LoadInfo`s. This matches how it unwraps generic messages from the command broker.