refactor(api): axis and mount types for ot3
The OT3 has a different set of axes and mounts, and it's good not to
confuse them with the OT2. While for now the upwards interface is the
same, by using these types in the api<->controller interface we can
eliminate string roundtrips. They percolate down to the hardware package
for use as keys in the motion math position dictionaries.

There's also the introduction of the new concept of "Axis kind". This is
sort of a generalization of the old Axis.gantry_axes methods. Especially
as we start having grippers and other tools, we have a lot of different
things that fall under the category of "plunger axis" or "z axis", and
it's nice to be able to refer to them generically. One such place where
it's nice to refer to them generically is in axis constraints and config
parsing, where we really just want a setting that applies to _any_ Z
axis.