feat(hardware): add network submodule with probe
Probe can be used to detect nodes present on the network by sending a
get status request ping and waiting for responses. This can be delicate
because you don't necessarily know when to stop listening, and we don't
want it to take forever; therefore, you can a) set a timeout and b)
provide a set of expected nodes, and the probe will terminate when it
either finds those nodes or hits the timeout, whichever happens first.

This still leaves some holes - if there are more nodes present than are
in the expected list, and the expected ones respond first, you won't see
them - but with proper use of the expected list it should be reliable.