refactor(hardware): allow fewer nodes in moves
We don't want to send move commands to nodes that aren't on the network.
If the higher layers have identified the set of nodes that are present,
they can send over that set of nodes and only move commands for the
specified nodes will be sent. To preserve test usability, this is
optional, and if no nodes are specified moves will still be sent to all
non-bootloader nodes.