refactor(api): Remove pipette pairing (#9399)
Pipette pairing was an experimental feature developed for making the
OT-2 higher throughput by simultaneously executing liquid handling
commands on both pipettes. The problem is that this requires the
pipettes to be extremely exactly physicall positioned relative to each other,
which basically isn't possible - they have to be exactly 36mm apart to
comply with the pitch of tipracks and ANSI wells, and that isn't
consistent enough. In the meantime, supporting the feature adds overhead
and complexity to every liquid handling code path.

This commit removes the (undocumented) feature from the protocol API and the hardware control API.