refactor(api): ot3: limit move command targets (#9408)
* feat(hardware): add network submodule with probe

Probe can be used to detect nodes present on the network by sending a
get status request ping and waiting for responses. This can be delicate
because you don't necessarily know when to stop listening, and we don't
want it to take forever; therefore, you can a) set a timeout and b)
provide a set of expected nodes, and the probe will terminate when it
either finds those nodes or hits the timeout, whichever happens first.

This still leaves some holes - if there are more nodes present than are
in the expected list, and the expected ones respond first, you won't see
them - but with proper use of the expected list it should be reliable.

* refactor(hardware): allow fewer nodes in moves

We don't want to send move commands to nodes that aren't on the network.
If the higher layers have identified the set of nodes that are present,
they can send over that set of nodes and only move commands for the
specified nodes will be sent. To preserve test usability, this is
optional, and if no nodes are specified moves will still be sent to all
non-bootloader nodes.

* refactor(api): ot3: Only specify move present nodes

We can probe the canbus network to identify which nodes are
present (this is simulated from the instrument spec, and for now in the
real thing is based on the instrument spec as well) and then pass only
those nodes into the move system, so we don't end up waiting forever on
nodes that aren't present.