refactor(api): Use ot3axis and mount internally
So it's nice to be able to use the new types to talk to the backends but
if you have to switch them from the old ones as inputs, and to the old
ones as outputs, you're still really just washing data around for no
gain. There's not much we can do about the outputs unless OT3Axis
contravaries with Axis and OT3Mount contravaries with Mount - but,
Union[OT3Axis, Axis] covaries with Axis, and Union[OT3Mount, Mount]
covaries with mount, so we can add the option for callers to use the
new, more-capable enums when they interact with an ot3 hardware control
API.

This leaves us in an interesting conundrum, though. We want to use them
internally; we want to have them as inputs; and we want to share
functionality.

Luckily, the enums are plenty generic to be used as generics, since
they're only being used as keys, especially if you can retrieve the type
objects in the rare instances where necessary from preconfiguration. So
InstrumentHandler, the major shared-functionality point, gets to be a
generic, and its little worker dataclasses get to be generics, and
because you don't have type maps or anything you get a bunch of
overloads to indicate that Axis-OT3Axis and Mount-OT3Mount are linked,
and there's a whole bunch of code with the result that things nicely
vary.

The second problem though is that now, if you have InstrumentHandler
nice and generic, then OT3API can't inherit from it but switch out all
its methods to union types because that doesn't work. So now it's a
has-a relationship. Wheeeeee