fix(hardware): network_probe functionality
Network probe was using get_status_request, which nothing has
implemented, and therefore nothing responds to. That's bad. Use
GetDeviceInfo instead which I think was what I was searching for in my
mind's brain originally.

Additionally, head responds to GetDeviceInfo, not head_l or head_r, so
handle that in the controller backends.