sometimes you just get a zero-length move
Because higher layers really want everything to be idempotent, they'll
always issue a command to move to a destination even if we happen to
already be at that destination, which is fine. we just need to not do
anything when we get that request (the motor controller handles this on
the ot2).