refactor(hardware): Parametrize motion axis types
This code previously used string axis names, which was fine, but it's
annoying to make it work with higher level code especially once that
higher level code transitions to axis enums.

The thing is, there's not really anything that this code _needs_ out of
any given axis name type. In 99% of cases it was just used as a dict
key, and the other cases were because of this Coordinates dataclass.

By making the Coordinates dataclass a generic type alias for a Mapping,
there's very few code changes needed (outside of tests, of course) to
make this code completely adaptable to basically any hashable type that
you feel like using as a key. Tests can use strings because it's easy;
and the api subpackage can use Axis, or OT3Axis, and all the typechecks
make it work.