refactor(hardware): move canbus bindings here
Take the canbus bindings from https://github.com/opentrons/ot3-firmware
python/ subdirectory and put them in opentrons_hardware.

We previously had a python package called opentrons_ot3_firmware that
was defined in a subdirectory of
https://github.com/opentrons/ot3-firmware. That repo contained basically
nothing but structures and enums that defined the structure and values
of canbus messages, so that they could be used from python and generated
to c and c++ from the same source. This was a pretty good idea! But it
fell down in the details of the monorepo package management solution
really not working well with things like poetry, or pyproject.toml, or
dealing with updates, which required elaborate workaround workflows
involving make teardown and make setup.

Instead, let's make the manual step more explicit and straightforward.
The python lives here, which means there's no manual upkeep in this
repository, and no weird workarounds in dependent packages for
optionally-present data. The workaround lives in ot3-firmware, where
when a developer wants to update canbus constant structures they now
make sure the monorepo with their changes is locally checked out and
then run a cmake command to generate the new header structure, which
they check in.