refactor(api): axis and mounts for ot3 (#9433)
* refactor(hardware): Parametrize motion axis types

This code previously used string axis names, which was fine, but it's
annoying to make it work with higher level code especially once that
higher level code transitions to axis enums.

The thing is, there's not really anything that this code _needs_ out of
any given axis name type. In 99% of cases it was just used as a dict
key, and the other cases were because of this Coordinates dataclass.

By making the Coordinates dataclass a generic type alias for a Mapping,
there's very few code changes needed (outside of tests, of course) to
make this code completely adaptable to basically any hashable type that
you feel like using as a key. Tests can use strings because it's easy;
and the api subpackage can use Axis, or OT3Axis, and all the typechecks
make it work.

* refactor(api): axis and mount types for ot3

The OT3 has a different set of axes and mounts, and it's good not to
confuse them with the OT2. While for now the upwards interface is the
same, by using these types in the api<->controller interface we can
eliminate string roundtrips. They percolate down to the hardware package
for use as keys in the motion math position dictionaries.

There's also the introduction of the new concept of "Axis kind". This is
sort of a generalization of the old Axis.gantry_axes methods. Especially
as we start having grippers and other tools, we have a lot of different
things that fall under the category of "plunger axis" or "z axis", and
it's nice to be able to refer to them generically. One such place where
it's nice to refer to them generically is in axis constraints and config
parsing, where we really just want a setting that applies to _any_ Z
axis.

* refactor(api): Use ot3axis and mount internally

So it's nice to be able to use the new types to talk to the backends but
if you have to switch them from the old ones as inputs, and to the old
ones as outputs, you're still really just washing data around for no
gain. There's not much we can do about the outputs unless OT3Axis
contravaries with Axis and OT3Mount contravaries with Mount - but,
Union[OT3Axis, Axis] covaries with Axis, and Union[OT3Mount, Mount]
covaries with mount, so we can add the option for callers to use the
new, more-capable enums when they interact with an ot3 hardware control
API.

This leaves us in an interesting conundrum, though. We want to use them
internally; we want to have them as inputs; and we want to share
functionality.

Luckily, the enums are plenty generic to be used as generics, since
they're only being used as keys, especially if you can retrieve the type
objects in the rare instances where necessary from preconfiguration. So
InstrumentHandler, the major shared-functionality point, gets to be a
generic, and its little worker dataclasses get to be generics, and
because you don't have type maps or anything you get a bunch of
overloads to indicate that Axis-OT3Axis and Mount-OT3Mount are linked,
and there's a whole bunch of code with the result that things nicely
vary.

The second problem though is that now, if you have InstrumentHandler
nice and generic, then OT3API can't inherit from it but switch out all
its methods to union types because that doesn't work. So now it's a
has-a relationship. Wheeeeee

* Use some extra type aliases