fix(api): make opentrons_hardware not required (#9605)
The OT2 doesn't have opentrons_hardware installed - that's only an OT3
thing. We have a responsibility to make that actually function. This
setup wasn't actually being tested.

This commit adds
- New makefile targets to "setup-ot2" in relevant python subprojects by
running a pipenv sync and then removing opentrons_hardware from the
virtualenv. This is not good, but absent being able to have separate
pipfiles it seems like the best fix available.
- Github workflow entries to lint and test under these conditions. This
is also not good because it will make these tests take longer, but same
thing about the conditions.

To be truly optional, it must be possible to import opentrons without
the ot3 feature flag set and not ever need opentrons_hardware. That was
mostly true, but there were some exceptions (default re-exports) that
needed to be updated.