refactor(hardware): can_mon resets styles
As with a lot of scripts that do terminal text styling, if you interrupt
can_mon or quit it in a surprising place, it can leave your terminal
styling overridden. Fix this by
- Having a more in-depth styling interface that always remembers to
reset terminal styles after each line
- Adding an atexit handler that resets terminal styling just in case

While we're at it, make the output pipe configurable which could let you
log stuff to a file (without the styling).