refactor(api): OT3 max instrument height is home
The OT2 needed a bunch of annoying workarounds and config values to
determine the highest position the pipette could achieve, for checking
if it could clear a given piece of labware. On the OT3, this highest
position is just the home position of the effector, and doesn't require
extra config.