feat(app): add managed file system storage for protocols (#9703)
create app-shell and app machinery to allow for files to be selected from the users filesystem and
copied into a managed directory within the App's userData folder.  This upload event will also
include creation of a unique identifier (protocolKey), a place to store the output of various
analyses run on this source file, and a nested src directory within which the source files will be
stored.  Wire up this new storage backend with Create, Read, and Delete functionality to the new
Protocol Landing page.

Closes #9483, Closes #9405

Co-authored-by: Sarah Breen <sarah.breen@opentrons.com>