feat(hardware): ot3: add tool info detector
We currently can detect what the head says is attached to it, but that
doesn't give us a ton of details. Now that we can query pipette
information from the pipettes directly, we can chain
- Asking what the head has
- Asking pipettes that the head says should be attached what they are

While properly only waiting for the ones that are attached to answer.

We will want to make a reactive version of this when we do the hardware
controller's attached tools being reactive, but to avoid extra churn the
first time through this is just a oneshot.