feat(hardware): ot3: add pipette querying (#9730)
- ot3: add pipette info messages
These messages can tell the host what kind of pipette is attached.

- ot3: add tool info detector
We currently can detect what the head says is attached to it, but that
doesn't give us a ton of details. Now that we can query pipette
information from the pipettes directly, we can chain
- Asking what the head has
- Asking pipettes that the head says should be attached what they are

While properly only waiting for the ones that are attached to answer.

We will want to make a reactive version of this when we do the hardware
controller's attached tools being reactive, but to avoid extra churn the
first time through this is just a oneshot.

- refactor(api): ot3: detect attached tools

Based on the new tool info detector, we can actually implement
get_attached_instruments() by querying the canbus. In many ways, this
duplicates the functionality of probe_network(); this now moves into the
realm of "debug tool" and perhaps should be removed.

- refactor(api): factor out probe filtering

- refactor(api): ot3: load gantry load from instrs
This will need to be updated to properly decide between low and high
throughput instruments and account for the gripper, but for now it
should work.