chore(build): Use pull request triggers for py (#9771)
Use pull request triggers as the primary mechanism for testing pull
requests, rather than push triggers.

Github workflows have a characteristically very low limit on the number
of concurrent MacOS jobs, no matter the subscription level, and has a
limit of 50 non-macOS jobs. Combined with the number of jobs that we
run, this greatly limits the velocity of merging and checking code.

However, the number of jobs that we run often represents duplicate work
- our test/lint workflows run on both pull_request (targeting the
predicted merge branch that github automatically creates) and on
push (targeting whatever obj/ref was just pushed). By using pull_request
triggers and limiting push to release, we can greatly cut down on the
amount of work we do, while still testing the same code.

The hole that's left is running tests on branches that developers have
pushed but do not want to open a pull request for, or that has a merge
conflict that prevents pull request checks from automatically running.
In this case, we have workflow_dispatch triggers so that the checks can
be run through the web UI, and we should probably add magic-comment
triggers for a similar thing.