refactor(api, hardware): ot3: Handle move position response (#9782)
* feat(hardware): return positions from moves

The MoveGroupRunner can now accumulate and return the current position
elements (which are now in micrometers, and are transformed to
millimeters in the process) from the remote hardware, providing a
consistent position estimate element to the python side of things.

* refactor(api): use move position responses

The move group runner now returns the position from the firmware after
it moves. Use that value set to update position rather than the
projected position based on the move group math.

* chore(hardware): add integration tests for moves

* fix integration tests

* fix home present-nodes limitation

This was causing exceptions because the present-axis checks were
mismatched and would present the move group runner with empty move
groups which it does NOT want to deal with. Avoid this with a new test
that makes sure that home doesn't generate empty move groups.

* fix(hardware): Move group handles more empty moves

The move groups are a data structure that can be "empty" on many
different levels; make sure we handle all of them.