refactor(api): add OT3API.capacitive_probe
This is a hardware control layer implementation of OT3 capacitive
probing, using new capabilities in opentrons_hardware to implement it.
The new function uses a type that will be in the config to configure
itself. It has similar semantics to aspirate() and dispense() - the
gantry will not move and it's assumed something else has put it in the
right place, and so on. The new method will move down until the
capacitive sensor triggers and then return the position at which it
triggered in deck coordinates before returning to a safe height.

Tests for the function only exist on the API layer because the
controller is a oneliner.

api test fixes and format