refactor(api-client, react-api-client, app): separate analyses from protocol resource
The protocol resource returned from the robot server's http api inlcudes analysis summaries in place
of full weight analysis record for performance reaso ns. This updates the client-side structures
that relied on the old resource. In order to support the new structure, this adds a new function and
hook to t he api-client and react-api-client respectively, for fetching the full analyses associated
with a protocol record.  In order to make these changes, much o f the legacy 'run' and 'upload'
pages of the app would have had to be updated. Because these views are slated for immediate removal
ahead of 5.1.0's release, they have been stubbed and removal has begun where functionality in the
new app is sufficient on its own.