refactor(hardware): Exact types for payloads (#10261)
Payloads had been specified in message dataclasses as their supertype,
which sort of discards a bunch of information. They should get hoisted
to their common supertype if you get a union of messages anyway, and
this way we can check any given message instantiation by the payload if
we want to - it certainly doesn't hurt anything.

It would be nice to have a typesafe way of getting the payload of a
message; in C++ this would be easy but mypy doesn't really have type
operations, so the only way to do it would be a giant explicit
overloaded function.