fix(hardware): instability in triangle move checks
Equal-leg triangle moves can occasionally, because of numerical
instability, appear as if the sum of the distances of their accelerating
and decelerating blocks is greater than the distance of the move. This
triggers a check intended to handle extremely-unequal triangle moves,
which cannot handle the case where the overall initial and final speeds
are almost exactly equal. In that case, the accel/decel legs have their
distance set to 0, but not their speeds, accelerations, or distances,
and the coast phase takes the entire difference. That causes the move to
go twice as far as it should.

This is fixed pretty easily by using a sloppier comparison to catch
issues of numerical stability.