fix(hardware): Fix equal leg triangle move handling (#10335)
* chore(hardware): add failing test for issue

When you try and internally blend a move that has the same starting and
ending velocity and has certain numerically-coincidental constraint and
distance values, it will sometimes accidentally trigger a part of the
planning code that is supposed to handle very unequal triangle moves,
and that absolutely does not handle the case where the initial and final
velocities are the same, and halfway ignores the accelerating and
decelerating legs of the move.

* fix(hardware): instability in triangle move checks

Equal-leg triangle moves can occasionally, because of numerical
instability, appear as if the sum of the distances of their accelerating
and decelerating blocks is greater than the distance of the move. This
triggers a check intended to handle extremely-unequal triangle moves,
which cannot handle the case where the overall initial and final speeds
are almost exactly equal. In that case, the accel/decel legs have their
distance set to 0, but not their speeds, accelerations, or distances,
and the coast phase takes the entire difference. That causes the move to
go twice as far as it should.

This is fixed pretty easily by using a sloppier comparison to catch
issues of numerical stability.